uki-util: move functions for inspecting kernel to uki-util.[ch]
